Refactoring, added ConversionUtil and MailService, moved conversion logic and mailsending logic out of main class
